reorganized user folders, uses index->routes->subfolders structure
