+++ collabs/paul/paul.js
@@ -1,0 +1,3 @@
+module.exports = function(req,res){
+    res.end('hello from Phil(?) at '+Date())
+}
